Move [simd.math] and add an implementation sketch
The <cmath> overloads want to call simd clones of the scalar math
functions. However, that's only possible when __FAST_MATH__ was defined
*before* including <math.h>.

Without __FAST_MATH__ we should dispatch to a function in the
shared/static library. The current _TargetTraits tag is too coarse,
though and needs to be replaced with something narrower.

include/ChangeLog:

	* bits/simd_details.h (_ArchFlags::_M_eval_as_f32): New.
	* bits/simd_math.h: Moved from simd_vec.h.
	(_GLIBCXX_SIMD_MATH_CALL, _GLIBCXX_SIMD_MATH_CALL2): New.
	* bits/simd_vec.h (_M_get, _M_get_low, _M_get_high): New.
	(_S_init): Use const-ref for arguments.
	([simd.math] [simd.complex.math]): Move to simd_math.h.
	* simd: Include bits/simd_math.h.